Add files via upload
An update to Sidebar.jsx that adds functionality.  Changes are:
Servers are now pulled from a user account (static currently, will update when authentication is in place)

The button at the bottom of the sidebar gives an option between:
      Create a new server - Creates a new server in the database and assigns a random icon.  Adds the server to the user's account for sidebar display.
                                         Allows the creator to choose the server ID.  Checks that the server ID chosen isn't already in use.
      Join a currently existing server - Inputting a serverId adds it to the user's database entry for display on the sidebar.
                                                          Checks that the ID entered is a valid server ID in the database.

A context menu has been added when right clicking on a server.  It currently has the option to update the Server's Name and leave the server, which unregisters it from the User's account to remove it from the sidebar.

The jsx styles have been split into a separate file and imported for clarity.