Enhancements to Chat Messages Component and Bug Fixes
Fixing bugs and enhancing the performance of the search components and file display. In the previous version, the files were successfully sent to Firebase but did not appear in the chatbox. I found it challenging to resolve this issue due to the complexity of the Chat Messages Component. So, I created a new Highlighted Text Component and integrated it into the Chat Messages Component. The warnings that appear come from the comments.